backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
cors==1.0.1
minio==7.2.0
python-multipart==0.0.6
pytesseract==0.3.10
PyMuPDF==1.23.8
Pillow>=8.0.0
sentence-transformers==3.0.1
pymilvus==2.3.7
tiktoken==0.5.2
numpy==1.24.3
huggingface-hub>=0.21.0
requests==2.31.0
